refactor update count command
