+++ liquibase-integration-tests/src/test/resources/liquibase/extension/testing/command/updateCount.test.groovy
@@ -48,7 +48,8 @@ Optional Args:
         ]
 
         expectedResults = [
-                statusCode   : 0
+                statusCode   : 0,
+                defaultChangeExecListener: 'not_null'
         ]
     }
 
@@ -63,7 +64,8 @@ Optional Args:
         ]
 
         expectedResults = [
-                statusCode   : 0
+                statusCode   : 0,
+                defaultChangeExecListener: 'not_null'
         ]
     }
 
@@ -80,7 +82,8 @@ Optional Args:
         ]
 
         expectedResults = [
-                statusCode   : 0
+                statusCode   : 0,
+                defaultChangeExecListener: 'not_null'
         ]
 
         expectedUI = [
@@ -114,7 +117,8 @@ DBMS mismatch:                1
         ]
 
         expectedResults = [
-                statusCode   : 0
+                statusCode   : 0,
+                defaultChangeExecListener: 'not_null'
         ]
 
         expectedUI = [
